backend/requirements.txt
bidict==0.23.1
blinker==1.9.0
certifi==2026.2.25
charset-normalizer==3.4.7
click==8.3.2
colorama==0.4.6
dnspython==2.8.0
eventlet==0.41.0
Flask==3.1.3
flask-cors==6.0.2
Flask-SocketIO==5.6.1
greenlet==3.4.0
h11==0.16.0
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
pymongo==4.16.0
PyJWT==2.8.0
python-dotenv==1.2.2
python-engineio==4.13.1
python-socketio==5.16.1
requests==2.33.1
simple-websocket==1.1.0
urllib3==2.6.3
Werkzeug==3.1.8
werkzeug==3.1.8
wsproto==1.3.2
